requirements.txt
Django>=5.1,<5.3
gunicorn>=23.0.0
whitenoise>=6.8.2
Pillow>=10.4.0
dj-database-url>=2.2.0
psycopg[binary]>=3.2.1
stripe>=11.6.0
python-dotenv>=1.0.1
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
